Frontend child registration and filtering contains USA related values, main page rendering nothing when doesn't have child (child registration instruction) after adding filter function fixed(useRef changed to using useState) Backend usa vaccination guide seed data added reference from USA CDC(https://www.cdc.gov/vaccines/hcp/imz-schedules/child-adolescent-age.html), migration for vaccination table data type changed from tiny int to int

+++ backend/src/migrations/CreateVaccineTable.ts
@@ -31,7 +31,7 @@ export class CreateVaccineTable20240806173026 implements MigrationInterface {
                     },
                     {
                         name: 'vaccine_minimum_recommend_date',
-                        type: 'tinyint',
+                        type: 'int',
                     },
                     {
                         name: 'vaccine_maximum_period_type',
@@ -40,7 +40,7 @@ export class CreateVaccineTable20240806173026 implements MigrationInterface {
                     },
                     {
                         name: 'vaccine_maximum_recommend_date',
-                        type: 'tinyint',
+                        type: 'int',
                         isNullable: true,
                     },
                     {
